+++ src/PockyBot.NET/PockyBotFactory.cs
@@ -43,9 +43,12 @@ namespace PockyBot.NET
                 new Status(pockyUserRepository, configRepository, pegHelper, loggerFactory.CreateLogger<Status>()),
                 new Finish(pockyUserRepository, pegResultsHelper, resultsUploader, directResultsMessageSender, loggerFactory.CreateLogger<Finish>()),
                 new Reset(pegRepository, loggerFactory.CreateLogger<Reset>()),
+                new Keywords(configRepository),
                 new Rotation(configRepository),
                 new LocationConfig(locationRepository, pockyUserRepository),
                 new UserLocation(pockyUserRepository, userLocationGetter, userLocationSetter, userLocationDeleter),
+                new StringConfig(configRepository),
+                new RemoveUser(pockyUserRepository, loggerFactory.CreateLogger<RemoveUser>()),
                 new Default(wrappedSettings)
             };