+++ src/PockyBot.NET/Services/Triggers/Help.cs
@@ -50,8 +50,8 @@ namespace PockyBot.NET.Services.Triggers
                     return CreatePegHelpMessage();
                 case Commands.Status:
                     return CreateStatusHelpMessage();
-                // case Commands.Keywords:
-                //     return CreateKeywordsHelpMessage();
+                case Commands.Keywords:
+                    return CreateKeywordsHelpMessage();
                 case Commands.Ping:
                     return CreatePingHelpMessage();
                 case Commands.Welcome:
@@ -70,16 +70,16 @@ namespace PockyBot.NET.Services.Triggers
                     return CreateFinishHelpMessage(user);
                 // case Commands.NumberConfig:
                 //     return CreateNumberConfigHelpMessage(user);
-                // case Commands.StringConfig:
-                //     return CreateStringConfigHelpMessage(user);
+                case Commands.StringConfig:
+                    return CreateStringConfigHelpMessage(user);
                 // case Commands.RoleConfig:
                 //     return CreateRoleConfigHelpMessage(user);
                 case Commands.LocationConfig:
                     return CreateLocationConfigHelpMessage(user);
                 case Commands.UserLocation:
                     return CreateUserLocationHelpMessage(user);
-                // case Commands.RemoveUser:
-                //     return CreateRemoveUserHelpMessage(user);
+                case Commands.RemoveUser:
+                    return CreateRemoveUserHelpMessage(user);
                 // case Commands.LocationWeight:
                 //     return CreateLocationWeightHelpMessage(user);
                 default:
@@ -97,7 +97,7 @@ namespace PockyBot.NET.Services.Triggers
             var newMessage = "## What I can do (List of Commands)\n\n" +
                  $"* {Commands.Peg}\n" +
                  $"* {Commands.Status}\n" +
-                 //$"* {Commands.Keywords}\n" +
+                 $"* {Commands.Keywords}\n" +
                  $"* {Commands.Ping}\n" +
                  $"* {Commands.Welcome}\n" +
                  $"* {Commands.Rotation}\n" +
@@ -125,16 +125,16 @@ namespace PockyBot.NET.Services.Triggers
                 newMessage += $"* {Commands.Finish}\n";
             }
 
-            // if (HasPermission(user, new []{Roles.Admin, Roles.Config})) {
-            //     newMessage += $"* {Commands.NumberConfig}\n";
-            //     newMessage += $"* {Commands.StringConfig}\n";
-            //     newMessage += $"* {Commands.RoleConfig}\n";
-            //     newMessage += $"* {Commands.LocationWeight}\n";
-            // }
+            if (HasPermission(user, new []{Roles.Admin, Roles.Config})) {
+                // newMessage += $"* {Commands.NumberConfig}\n";
+                newMessage += $"* {Commands.StringConfig}\n";
+                // newMessage += $"* {Commands.RoleConfig}\n";
+                // newMessage += $"* {Commands.LocationWeight}\n";
+            }
 
-            // if (HasPermission(user, new []{Roles.Admin, Roles.RemoveUser})) {
-            //     newMessage += $"* {Commands.RemoveUser}\n";
-            // }
+            if (HasPermission(user, new []{Roles.Admin, Roles.RemoveUser})) {
+                newMessage += $"* {Commands.RemoveUser}\n";
+            }
 
             newMessage += $"\nFor more information on a command type `@{_pockyBotSettings.BotName} help command-name` or direct message me with `help command-name`\n";
             newMessage += "\nI am still being worked on, so more features to come.";
@@ -160,12 +160,12 @@ namespace PockyBot.NET.Services.Triggers
                 "1. I will PM you number of pegs you have left and who you gave it to.";
         }
 
-        // private string CreateKeywordsHelpMessage()
-        // {
-        //     return "### How to check the available keywords 🔑!\n" +
-        //         $"1. To get a list of the available keywords, type: `@{_pockyBotSettings.BotName} {Commands.Keywords}` OR direct message me with `{Commands.Keywords}`.\n" +
-        //         "1. I will respond in the room you messaged me in with a list of keywords.";
-        // }
+        private string CreateKeywordsHelpMessage()
+        {
+            return "### How to check the available keywords 🔑!\n" +
+                $"1. To get a list of the available keywords, type: `@{_pockyBotSettings.BotName} {Commands.Keywords}` OR direct message me with `{Commands.Keywords}`.\n" +
+                "1. I will respond in the room you messaged me in with a list of keywords.";
+        }
 
         private string CreatePingHelpMessage()
         {
@@ -247,17 +247,17 @@ namespace PockyBot.NET.Services.Triggers
         //     }
         //     return CreateDefaultHelpMessage();
         // }
-        //
-        // private string CreateStringConfigHelpMessage(PockyUser user)
-        // {
-        //     if (HasPermission(user, new[] {Roles.Admin, Roles.Config})) {
-        //         return "### How to configure string config values 🎻!\n" +
-        //             $"1. To get/edit/refresh/delete string config values, type `@{_pockyBotSettings.BotName} {Commands.StringConfig} {Object.values(ConfigAction).join('|')} {{name}} {{value}}`\n" +
-        //             "1. I will respond in the room you messaged me in.";
-        //     }
-        //     return CreateDefaultHelpMessage();
-        // }
-        //
+
+        private string CreateStringConfigHelpMessage(PockyUser user)
+        {
+            if (HasPermission(user, new[] {Roles.Admin, Roles.Config})) {
+                return "### How to configure string config values 🎻!\n" +
+                    $"1. To get/add/delete string config values, type `@{_pockyBotSettings.BotName} {Commands.StringConfig} {string.Join("|",ConfigActions.All())} {{name}} {{value}}`\n" +
+                    "1. I will respond in the room you messaged me in.";
+            }
+            return CreateDefaultHelpMessage();
+        }
+
         // private string CreateRoleConfigHelpMessage(PockyUser user)
         // {
         //     if (HasPermission(user, new[] {Roles.Admin, Roles.Config})) {
@@ -298,16 +298,16 @@ namespace PockyBot.NET.Services.Triggers
                 "1. I will respond in the room you messaged me in.";
         }
 
-        // private string CreateRemoveUserHelpMessage(PockyUser user)
-        // {
-        //     if (HasPermission(user, new[] {Roles.Admin, Roles.RemoveUser})) {
-        //         return "### How to remove users 🛑!\n" +
-        //             $"1. To remove a user, type `@{_pockyBotSettings.BotName} {Commands.RemoveUser} {{@User}}|'{{username}}'`\n" +
-        //             "1. I will respond in the room you messaged me in.";
-        //     }
-        //     return CreateDefaultHelpMessage();
-        // }
-        //
+        private string CreateRemoveUserHelpMessage(PockyUser user)
+        {
+            if (HasPermission(user, new[] {Roles.Admin, Roles.RemoveUser})) {
+                return "### How to remove users 🛑!\n" +
+                    $"1. To remove a user, type `@{_pockyBotSettings.BotName} {Commands.RemoveUser} {{@User}}|'{{username}}'`\n" +
+                    "1. I will respond in the room you messaged me in.";
+            }
+            return CreateDefaultHelpMessage();
+        }
+
         // private string CreateLocationWeightHelpMessage(PockyUser user)
         // {
         //     if (HasPermission(user, new[] {Roles.Admin, Roles.Config})) {