+++ src/PockyBot.NET/StartupExtensions.cs
@@ -42,8 +42,11 @@ namespace PockyBot.NET
             services.AddTransient<ITrigger, Status>();
             services.AddTransient<ITrigger, Finish>();
             services.AddTransient<ITrigger, Reset>();
+            services.AddTransient<ITrigger, Keywords>();
             services.AddTransient<ITrigger, Rotation>();
             services.AddTransient<ITrigger, LocationConfig>();
+            services.AddTransient<ITrigger, StringConfig>();
+            services.AddTransient<ITrigger, RemoveUser>();
             services.AddTransient<ITrigger, Default>();
         }
     }