+++ CHANGELOG.md
@@ -1,3 +1,61 @@
+# [1.7.0](https://github.com/GlobalX/PockyBot.NET/compare/v1.6.2...v1.7.0) (2020-03-03)
+
+
+### Features
+
+* add keywords trigger ([b9e6bbd](https://github.com/GlobalX/PockyBot.NET/commit/b9e6bbd3117ca594a3cd52fc67555164ff9e5e9e))
+
+## [1.6.2](https://github.com/GlobalX/PockyBot.NET/compare/v1.6.1...v1.6.2) (2020-02-25)
+
+
+### Bug Fixes
+
+* fix db representation so that users can be accessed and removed properly ([4d2f153](https://github.com/GlobalX/PockyBot.NET/commit/4d2f15370591dc37805a0a361a2890d3d6ad95f1))
+
+## [1.6.1](https://github.com/GlobalX/PockyBot.NET/compare/v1.6.0...v1.6.1) (2020-02-21)
+
+
+### Bug Fixes
+
+* add removeuser to help trigger ([42a5970](https://github.com/GlobalX/PockyBot.NET/commit/42a5970d08a237857b23904f40a3ce0694b41676))
+* fix removeuser null reference and add logging ([3e6519c](https://github.com/GlobalX/PockyBot.NET/commit/3e6519c67c9dc5f3ad6259f5788e30eea6d1fcbb))
+
+# [1.6.0](https://github.com/GlobalX/PockyBot.NET/compare/v1.5.1...v1.6.0) (2020-02-17)
+
+
+### Bug Fixes
+
+* code style issues - make test objects readonly ([1334a30](https://github.com/GlobalX/PockyBot.NET/commit/1334a3058d5f11695073867ccd7d96bcb31dcbaa))
+
+
+### Features
+
+* add removeuser trigger ([07d9213](https://github.com/GlobalX/PockyBot.NET/commit/07d9213fcf0cbb9f4cb1b6628d9d31e1de79eb46))
+
+## [1.5.1](https://github.com/GlobalX/PockyBot.NET/compare/v1.5.0...v1.5.1) (2020-02-10)
+
+
+### Bug Fixes
+
+* update stringconfig help message ([8b9e635](https://github.com/GlobalX/PockyBot.NET/commit/8b9e63510d59bedde750cb2d750a240c54a9d893))
+
+# [1.5.0](https://github.com/GlobalX/PockyBot.NET/compare/v1.4.0...v1.5.0) (2020-02-10)
+
+
+### Bug Fixes
+
+* code style fixes ([a77086b](https://github.com/GlobalX/PockyBot.NET/commit/a77086b3b237871aa29614bb0c2741aa5278e708))
+* PR comments for string config ([20ef1cb](https://github.com/GlobalX/PockyBot.NET/commit/20ef1cb314f1ce453614655899082a530d55c6eb))
+
+
+### Features
+
+* add stringconfig to factory and startup extensions, and add unit tests ([e29cbfb](https://github.com/GlobalX/PockyBot.NET/commit/e29cbfbeaa13ddb282c5c19d029e1516f62ad86e))
+* add stringconfig to help and update unit tests ([8bff45f](https://github.com/GlobalX/PockyBot.NET/commit/8bff45f7a005bc3751532b117389e7ee6ef9dcea))
+* implement stringconfig add and delete ([d045921](https://github.com/GlobalX/PockyBot.NET/commit/d0459215a12137bfaa363e20744c40bf23029548))
+* implement stringconfig get ([e5bb292](https://github.com/GlobalX/PockyBot.NET/commit/e5bb2923463ea28e6df676c76da00c2c58e4232b))
+* update stringconfig to display configuration in a list rather than a table ([f772211](https://github.com/GlobalX/PockyBot.NET/commit/f772211fbf66a671018e9c467e42c3692522eb63))
+
 # [1.4.0](https://github.com/GlobalX/PockyBot.NET/compare/v1.3.1...v1.4.0) (2020-01-31)